rename: lib->utils 파일명 변경, hooks,types 폴더 추가
