+++ src/shared/utils/validation.ts
